bugFix: 오류 부분 수정 (JPA 연관관계)
