+++ backend/src/test/java/com/example/backend/social/feed/implement/FeedTestHelper.java
@@ -56,9 +56,6 @@ public class FeedTestHelper {
 	private CommentRepository commentRepository;
 
 	@Autowired
-	private FollowRepository followRepository;
-
-	@Autowired
 	private LikeService likeService;
 
 	@Autowired
